NutriSnap-AI/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
firebase-admin
google-generativeai
pillow
